move stats resource to chain slice
