+++ src/features/gas-tracker/pages/index/GasTracker.pw.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import * as statsMock from 'src/slices/home/mocks/stats';
+import * as statsMock from 'src/slices/chain/stats/mocks';
 
 import * as statsLineMock from 'src/features/chain-stats/mocks/line';
 import * as statsLinesMock from 'src/features/chain-stats/mocks/lines';
