+++ src/shell/top-bar/TopBar.pw.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import * as statsMock from 'src/slices/home/mocks/stats';
+import * as statsMock from 'src/slices/chain/stats/mocks';
 
 import * as csvExportStorageMock from 'src/features/csv-export/mocks/storage';
 import type { StorageItem } from 'src/features/csv-export/utils/storage';
