+++ src/slices/chain/stats/mocks.ts
