+++ src/slices/home/pages/index/stats/Stats.pw.tsx
@@ -2,8 +2,8 @@ import type { Locator } from '@playwright/test';
 import React from 'react';
 
 import * as blockListMock from 'src/slices/block/mocks/list';
+import * as statsMock from 'src/slices/chain/stats/mocks';
 import { HomeDataContextProvider } from 'src/slices/home/contexts/home-data-context';
-import * as statsMock from 'src/slices/home/mocks/stats';
 
 import { test, expect } from 'playwright/lib';
