+++ src/slices/tx/pages/index/stats/TxsStats.pw.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import * as statsMock from 'src/slices/home/mocks/stats';
+import * as statsMock from 'src/slices/chain/stats/mocks';
 import * as txsStatsMock from 'src/slices/tx/mocks/stats';
 
 import { test, expect } from 'playwright/lib';
